moved common patten implementations to patterns
