+++ package.json
@@ -7,7 +7,7 @@
     "clean": "rimraf ./packages/*/dist",
     "build": "tsc --build",
     "lint": "eslint .",
-    "pretest": "yarn lint && yarn build",
+    "pretest": "yarn build && yarn lint",
     "test": "yarn test:only",
     "test:only": "mocha \"packages/*/dist{,/cjs}/test/**/*.unit.js\"",
     "test:dev": "mocha -r @ts-tools/node/r \"packages/*/test/**/*.unit.ts{,x}\"",
