+++ packages/patterns/src/event-emitter.ts
