+++ packages/patterns/src/index.ts
@@ -1,1 +1,12 @@
+/**
+ * @packageDocumentation
+ * Useful data structures such as
+ * LRU-Cache, set-multi-map,
+ * Signal, EventEmitter, Disposable, Debouncer
+ */
 export * from './create-disposables';
+export * from './event-emitter';
+export * from './signal';
+export * from './lru-cache';
+export * from './set-multi-map';
+export * from './debouncer';
