+++ packages/patterns/src/lru-cache.ts
