+++ packages/patterns/src/signal.ts
